[Abhi]ADDED:Creating a note by using USERID

+++ Fundo_Notes/src/services/notes.service.js
@@ -7,22 +7,23 @@ export const createNote = async (body) => {
 };
 
 //get all notes
-export const getAllNotes = async () => {
-  const data = await Notes.find();
+export const getAllNotes = async (UserID) => {
+  const data = await Notes.find({UserID});
   return data;
 };
 
 //get a note by id
-export const getNote = async (id) => {
-  const data = await Notes.findById(id);
+export const getNote = async (id,UserID) => {
+  const data = await Notes.findById({ _id: id, UserID: UserID });
   return data;
 };
 
 //update a note
-export const updateNote = async (_id, body) => {
+export const updateNote = async (_id,UserID, body) => {
   const data = await Notes.findByIdAndUpdate(
     {
-      _id
+      _id: id,
+      UserID: UserID
     },
     body,
     {
@@ -33,14 +34,14 @@ export const updateNote = async (_id, body) => {
 };
 
 //delete a Single note
-export const deleteNote = async (id) => {
-  await Notes.findByIdAndDelete(id);
+export const deleteNote = async (id, UserID) => {
+  await Notes.findByIdAndDelete({ _id: id, UserID: UserID });
   return '';
 };
 
 //archieve a note
-export const archiveNote = async (_id) => {
-  const note = await Notes.findOne({ _id: _id });
+export const archiveNote = async (_id,UserID) => {
+  const note = await Notes.findOne({ _id: id, UserId: UserID });
   const isArchived = note.isArchived === false ? true : false;
   const data = await Notes.findByIdAndUpdate(
     {
@@ -55,8 +56,8 @@ export const archiveNote = async (_id) => {
 };
 
 //trash a note
-export const trashNote = async (_id) => {
-  const note = await Notes.findOne({ _id: _id });
+export const trashNote = async (_id,UserID) => {
+  const note = await Notes.findOne({ _id: id,UserID: UserID });
   const isTrash = note.isTrash === false ? true : false;
   const data = await Notes.findByIdAndUpdate(
     {
